+++ vpmobil/extensions/config_presets.py
@@ -2,7 +2,8 @@ import re
 
 StBennoGymnasium = {
     "AUFZÄHLUNGS_SEPARATOR": " ",
-    "KLASSENBEZEICHNER_PATTERN": r"(?P<stufe>[1-9][0-9]?)(?P<suffix>[a-z])"
+    "KLASSENBEZEICHNER_PATTERN": r"(?P<stufe>[1-9][0-9]?)(?P<suffix>[a-z])",
+    "STUNDE_HERVERLEGT_PATTERN": r"verlegt von St\.(?P<periode>\d+);"
 
 }