+++ documentation/anderes/changelog.md
@@ -9,6 +9,7 @@ icon: clock-rotate-left
 ### 🚀 Neue Funktionen
 
 * Eigenschaft `fachmeta` zu `Stunde` hinzugefügt, die die Information aus dem `Ku2`-Tag der Quelldaten bereitstellt
+* Eigenschaft `verlegt` zu `Stunde` hinzugefügt
 
 ### 🔧 Änderungen